@@ -14,8 +14,9 @@ define(function (require) {
       makeLabel: function () {
         let field = this.field();
         let format = (field && field.format) || fieldFormats.getDefaultInstance('number');
+        const label = this.params.customLabel || this.fieldDisplayName();
 
-        return 'Percentile rank ' + format.convert(this.key, 'text') + ' of "' + this.fieldDisplayName() + '"';
+        return 'Percentile rank ' + format.convert(this.key, 'text') + ' of "' + label + '"';
       }
     };
 
