@@ -1,7 +1,6 @@
 define(function (require) {
   let _ = require('lodash');
   let $ = require('jquery');
-  let addWordBreaks = require('ui/utils/add_word_breaks');
   let module = require('ui/modules').get('app/discover');
 
   require('ui/highlight');
@@ -150,18 +149,14 @@ define(function (require) {
         /**
          * Fill an element with the value of a field
          */
-        function _displayField(row, fieldName, breakWords) {
+        function _displayField(row, fieldName, truncate) {
           let indexPattern = $scope.indexPattern;
           let text = indexPattern.formatField(row, fieldName);
 
-          if (breakWords) {
-            text = addWordBreaks(text, MIN_LINE_LENGTH);
-
-            if (text.length > MIN_LINE_LENGTH) {
-              return truncateByHeightTemplate({
-                body: text
-              });
-            }
+          if (truncate && text.length > MIN_LINE_LENGTH) {
+            return truncateByHeightTemplate({
+              body: text
+            });
           }
 
           return text;
