@@ -23,16 +23,18 @@ define(function () {
      * @return {[type]} [description]
      */
     let getKnownKibanaTypes = _.once(function () {
-      let indexName = kbnIndex;
       return es.indices.getFieldMapping({
         // only concerned with types in this kibana index
-        index: indexName,
+        index: kbnIndex,
         // check all types
         type: '*',
         // limit the response to just the _source field for each index
         field: '_source'
       }).then(function (resp) {
-        return _.keys(resp[indexName].mappings);
+        // kbnIndex is not sufficient here, if the kibana indexed is aliased we need to use
+        // the root index name as key
+        const index = _.keys(resp)[0];
+        return _.keys(resp[index].mappings);
       });
     });
 
