@@ -11,8 +11,7 @@ let data = require('fixtures/vislib/mock_data/date_histogram/_series');
 describe('Vislib Dispatch Class Test Suite', function () {
 
   function destroyVis(vis) {
-    $(vis.el).remove();
-    vis = null;
+    vis.destroy();
   }
 
   function getEls(el, n, type) {
