@@ -41,8 +41,7 @@ dateHistogramArray.forEach(function (data, i) {
     }));
 
     afterEach(function () {
-      $(vis.el).remove();
-      vis = null;
+      vis.destroy();
     });
 
     describe('render Method', function () {
