@@ -13,7 +13,7 @@ define(function (require) {
       var settingsPage;
       var discoverPage;
       var visualizePage;
-      var remote;
+      //var remote;
 
       bdd.before(function () {
         common = new Common(this.remote);
@@ -21,7 +21,6 @@ define(function (require) {
         settingsPage = new SettingsPage(this.remote);
         discoverPage = new DiscoverPage(this.remote);
         visualizePage = new VisualizePage(this.remote);
-        remote = this.remote;
         var fromTime = '2015-09-19 06:31:44.000';
         var toTime = '2015-09-23 18:31:44.000';
 
@@ -88,32 +87,7 @@ define(function (require) {
         var testSubName = 'LineChart';
         var vizName1 = 'Visualization ' + testSubName;
 
-        bdd.it('should be able to save and load', function pageHeader() {
-
-          common.debug('Start of test' + testSubName + 'Visualization');
-          var remote = this.remote;
-
-          return visualizePage.saveVisualization(vizName1)
-          .then(function (message) {
-            common.debug('Saved viz message = ' + message);
-            expect(message).to.be('Visualization Editor: Saved Visualization \"' + vizName1 + '\"');
-          })
-          .then(function testVisualizeWaitForToastMessageGone() {
-            return visualizePage.waitForToastMessageGone();
-          })
-          .then(function () {
-            return visualizePage.loadSavedVisualization(vizName1);
-          })
-          .then(function waitForVisualization() {
-            return visualizePage.waitForVisualization();
-          })
-          .catch(common.handleError(this));
-        });
-
-
         bdd.it('should show correct chart, take screenshot', function pageHeader() {
-
-          var remote = this.remote;
 
           // this test only verifies the numerical part of this data
           // it could also check the legend to verify the extensions
@@ -142,10 +116,44 @@ define(function (require) {
           .catch(common.handleError(this));
         });
 
-        bdd.it('should show correct data', function pageHeader() {
+        bdd.it('should show correct chart order by Term', function pageHeader() {
 
-          var remote = this.remote;
-          var expectedChartData = ['jpg 9,109', 'css 2,159', 'png 1,373', 'gif 918', 'php 445'];
+          // this test only verifies the numerical part of this data
+          // https://github.com/elastic/kibana/issues/8141
+          var expectedChartData = ['png 1,373', 'php 445', 'jpg 9,109', 'gif 918', 'css 2,159'];
+
+          common.debug('Order By = Term');
+          return visualizePage.selectOrderBy('_term')
+          .then(function () {
+            return common.sleep(1000);
+          })
+          .then(function clickGo() {
+            return visualizePage.clickGo();
+          })
+          .then(function () {
+            return common.sleep(1000);
+          })
+          .then(function () {
+            return visualizePage.getLineChartData('fill="#57c17b"')
+            .then(function showData(data) {
+              common.debug('data=' + data);
+              common.saveScreenshot('Visualize-line-chart');
+              var tolerance = 10; // the y-axis scale is 10000 so 10 is 0.1%
+              for (var x = 0; x < data.length; x++) {
+                common.debug('x=' + x + ' expectedChartData[x].split(\' \')[1] = ' +
+                  (expectedChartData[x].split(' ')[1]).replace(',', '') + '  data[x]=' + data[x] +
+                  ' diff=' + Math.abs(expectedChartData[x].split(' ')[1].replace(',', '') - data[x]));
+                expect(Math.abs(expectedChartData[x].split(' ')[1].replace(',', '') - data[x]) < tolerance).to.be.ok();
+              }
+              common.debug('Done');
+            });
+          });
+        });
+
+
+        bdd.it('should show correct data, ordered by Term', function pageHeader() {
+
+          var expectedChartData = ['png 1,373', 'php 445', 'jpg 9,109', 'gif 918', 'css 2,159'];
 
           return visualizePage.collapseChart()
           .then(function getDataTableData() {
@@ -158,6 +166,26 @@ define(function (require) {
           .catch(common.handleError(this));
         });
 
+
+        bdd.it('should be able to save and load', function pageHeader() {
+
+          common.debug('Start of test' + testSubName + 'Visualization');
+          return visualizePage.saveVisualization(vizName1)
+          .then(function (message) {
+            common.debug('Saved viz message = ' + message);
+            expect(message).to.be('Visualization Editor: Saved Visualization \"' + vizName1 + '\"');
+          })
+          .then(function testVisualizeWaitForToastMessageGone() {
+            return visualizePage.waitForToastMessageGone();
+          })
+          .then(function () {
+            return visualizePage.loadSavedVisualization(vizName1);
+          })
+          .then(function waitForVisualization() {
+            return visualizePage.waitForVisualization();
+          })
+          .catch(common.handleError(this));
+        });
 
       });
     });
